requirements.txt
bcrypt==4.0.1
cffi==1.15.1
cryptography==40.0.2
et-xmlfile==1.1.0
future==0.18.3
netmiko==4.1.2
ntc-templates==3.3.0
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.1
paramiko==3.1.0
pycparser==2.21
PyNaCl==1.5.0
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
scp==0.14.5
six==1.16.0
tenacity==8.2.2
textfsm==1.1.2
tzdata==2023.3